+++ src/lib.rs
@@ -49,5 +49,6 @@
 
 mod num;
 
+pub use crate::num::CustomDisplay;
 pub use crate::num::Num;
 pub use crate::num::ParseNumError;