+++ tests/test_num.rs
@@ -171,6 +171,17 @@ fn num_precision_fill_zeros() {
 }
 
 #[test]
+fn num_minimum_precision() {
+  assert_eq!(format!("{}", Num::new(1, 2).display().min_precision(2)), "0.50");
+  assert_eq!(format!("{}", Num::new(1, 3).display().min_precision(2)), "0.33333333");
+  assert_eq!(format!("{}", Num::new(2, 3).display().min_precision(2)), "0.66666667");
+  assert_eq!(format!("{}", Num::new(1, 3).display()), "0.33333333");
+  assert_eq!(format!("{}", Num::new(2, 3).display()), "0.66666667");
+  assert_eq!(format!("{:.2}", Num::new(1, 3).display()), "0.33");
+  assert_eq!(format!("{:.2}", Num::new(2, 3).display()), "0.67");
+}
+
+#[test]
 fn num_round_zero() {
   let num = Num::from_str("0.0").unwrap().round();
   assert_eq!(num, Num::new(0, 1));