+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 Unreleased
 ----------
+- Added support for printing a `Num` with a minimum precision
 - Added export for `ParseNumError` error type